removal of expired hitchhikers works :)

+++ Hitchhicker-Endpoint-V1/Hitchhiker-Endpoint/Controllers/CreateArgs.cs
@@ -3,7 +3,7 @@
     public class CreateArgs : object
     {
         public string? Location { get; set; }
-        public int? MinutesTillDisposal { get; set; }
+        public double? MinutesTillDisposal { get; set; }
         public string? Destination { get; set; }
     }
 }
